Blood-Report-Disease-Diagnosis-App-main/requirements.txt
Flask==2.3.2
numpy==1.26
pandas==2.1.1
scikit-learn==1.4.2
google-generativeai>=0.3.0
Werkzeug>=2.0.0
Pillow>=9.0.0
